// src/app/platform-service-api/model/eSIRefEndpointSyncTracker.ts
/**
 * EveKit Frontend Server
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ESI reference endpoint synchronization tracker
 */
export interface ESIRefEndpointSyncTracker {
    refreshed?: boolean;
    /**
     * Uniquer tracker ID
     */
    tid?: number;
    /**
     * endpoint
     */
    endpoint?: ESIRefEndpointSyncTracker.EndpointEnum;
    /**
     * Scheduled tracker start time (milliseconds UTC)
     */
    scheduled?: number;
    /**
     * Tracker start time (milliseconds UTC)
     */
    syncStart?: number;
    /**
     * Tracker end time (milliseconds UTC)
     */
    syncEnd?: number;
    /**
     * status
     */
    status?: ESIRefEndpointSyncTracker.StatusEnum;
    /**
     * status detail message
     */
    detail?: string;
}
export namespace ESIRefEndpointSyncTracker {
    export type EndpointEnum = 'REF_SERVER_STATUS' | 'REF_ALLIANCE' | 'REF_SOV_MAP' | 'REF_SOV_CAMPAIGN' | 'REF_SOV_STRUCTURE' | 'REF_FW_WARS' | 'REF_FW_STATS' | 'REF_FW_SYSTEMS' | 'REF_FW_FACTION_LEADERBOARD' | 'REF_FW_CORP_LEADERBOARD' | 'REF_FW_CHAR_LEADERBOARD';
    export const EndpointEnum = {
        SERVERSTATUS: 'REF_SERVER_STATUS' as EndpointEnum,
        ALLIANCE: 'REF_ALLIANCE' as EndpointEnum,
        SOVMAP: 'REF_SOV_MAP' as EndpointEnum,
        SOVCAMPAIGN: 'REF_SOV_CAMPAIGN' as EndpointEnum,
        SOVSTRUCTURE: 'REF_SOV_STRUCTURE' as EndpointEnum,
        FWWARS: 'REF_FW_WARS' as EndpointEnum,
        FWSTATS: 'REF_FW_STATS' as EndpointEnum,
        FWSYSTEMS: 'REF_FW_SYSTEMS' as EndpointEnum,
        FWFACTIONLEADERBOARD: 'REF_FW_FACTION_LEADERBOARD' as EndpointEnum,
        FWCORPLEADERBOARD: 'REF_FW_CORP_LEADERBOARD' as EndpointEnum,
        FWCHARLEADERBOARD: 'REF_FW_CHAR_LEADERBOARD' as EndpointEnum
    }
    export type StatusEnum = 'NOT_PROCESSED' | 'FINISHED' | 'ERROR' | 'WARNING';
    export const StatusEnum = {
        NOTPROCESSED: 'NOT_PROCESSED' as StatusEnum,
        FINISHED: 'FINISHED' as StatusEnum,
        ERROR: 'ERROR' as StatusEnum,
        WARNING: 'WARNING' as StatusEnum
    }
}